requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.4
gunicorn==21.2.0
numpy==2.1.2
scipy==1.14.1
networkx==3.3
setuptools>=65
# scikit-fuzzy 0.4.2 no es compatible con Python 3.13 (usa 'imp'). Usar versión de GitHub.
scikit-fuzzy @ git+https://github.com/scikit-fuzzy/scikit-fuzzy.git